fix(spotlight): adapter 進不了 image，線上 117 課 lesson_content 全為 null
get_lesson_content() 動態 import scripts/spotlight_to_lesson_content.py，
但 backend/Dockerfile 只 COPY app/ data/ alembic/ —— scripts/ 從未被複製，
也複製不到（build context 是 ./backend，scripts/ 在上一層）。

容器內 → No module named 'spotlight_to_lesson_content' → fail-closed 回 null。
124 課有素材，只有 7 課（早期手工策展 dev7，走不需 adapter 的路徑）有值。

四個月沒被發現：錯誤被 fail-closed 吃掉（不噴 500、health 200、CI 綠），
而本機一切正常（scripts/ 就在旁邊）。既有 122 條測試在修之前也全綠。

改動：
1. adapter 搬進 backend/app/services/ —— 它本來就 from app.schemas /
   from app.services，是 runtime 依賴不是 script。COPY app/ 自動帶進 image。
2. scripts/ 留別名模組（sys.modules[__name__] = _real）—— CLI、dry-run
   與既有測試照常運作，含私有函式存取。re-export 不複製，不會漂移。
3. 路徑錨定改 _BACKEND_ROOT，容器內外皆正確。
4. loader 改走正常 package import。

容器內驗證（同一條呼叫鏈）：
  prod 現行 image  → 產出成功 7 / 失敗 117
  本 PR 建的 image → 產出成功 124 / 失敗 0
本機 pytest 122 passed（但本機綠不算證據，這個 bug 就是本機綠線上壞）。

catalog/manifest.json 未更動 → 線上顯示行為不變，不會點亮任何新課。
本 PR 只修機制，點亮與否是另一個決定。
